Make some libraries optional
perl and libzeep are only required by mkhssp, so a warning is given in the
configure script instead of an error. This fixes #66 and #65.

bz2 is also optional and is controlled by a preprocessor flag. If the system
doesn't have the library installed, compressed files will not be supported.
This fixes issue #58.